STYLE: clang-tidy bugprone-implicit-widening-of-multiplication-result
The check diagnoses instances where a result of a multiplication is implicitly
widened, and suggests (with fix-it) to either silence the code by making
widening explicit, or to perform the multiplication in a wider type, to avoid
the widening afterwards.

This is fine in general, but if width * height overflows, you end up
wrapping back to the beginning of base instead of processing the entire
requested buffer.

Indeed, this only matters for pretty large buffers (4GB+), but that can
happen very easily for example in image processing, where for that to
happen you “only” need a ~269MPix image.

+++ Modules/Core/Common/include/itkSymmetricEigenAnalysis.hxx
@@ -18,6 +18,8 @@
 #ifndef itkSymmetricEigenAnalysis_hxx
 #define itkSymmetricEigenAnalysis_hxx
 
+#include <cstddef>
+
 #include "itkMath.h"
 #include "itkMakeUniqueForOverwrite.h"
 
@@ -60,7 +62,7 @@ SymmetricEigenAnalysis<TMatrix, TVector, TEigenMatrix>::ComputeEigenValuesLegacy
   const auto workArea1 = std::make_unique<double[]>(m_Dimension);
 
   // Copy the input matrix
-  const auto inputMatrix = make_unique_for_overwrite<double[]>(m_Dimension * m_Dimension);
+  const auto inputMatrix = make_unique_for_overwrite<double[]>(static_cast<size_t>(m_Dimension * m_Dimension));
   const auto dVector = make_unique_for_overwrite<double[]>(m_Dimension);
 
   unsigned int k = 0;
@@ -94,10 +96,10 @@ SymmetricEigenAnalysis<TMatrix, TVector, TEigenMatrix>::ComputeEigenValuesAndVec
   TEigenMatrix &  EigenVectors) const
 {
   const auto workArea1 = std::make_unique<double[]>(m_Dimension);
-  const auto workArea2 = std::make_unique<double[]>(m_Dimension * m_Dimension);
+  const auto workArea2 = std::make_unique<double[]>(static_cast<size_t>(m_Dimension * m_Dimension));
 
   // Copy the input matrix
-  const auto inputMatrix = make_unique_for_overwrite<double[]>(m_Dimension * m_Dimension);
+  const auto inputMatrix = make_unique_for_overwrite<double[]>(static_cast<size_t>(m_Dimension * m_Dimension));
   const auto dVector = make_unique_for_overwrite<double[]>(m_Dimension);
 
   unsigned int k = 0;
