+++ Modules/Core/QuadEdgeMesh/test/itkQuadEdgeMeshAddFaceTest2.cxx
@@ -16,6 +16,8 @@
  *
  *=========================================================================*/
 
+#include <cstddef>
+
 #include "itkQuadEdgeMeshTopologyChecker.h"
 
 
@@ -92,7 +94,7 @@ itkQuadEdgeMeshAddFaceTest2(int, char *[])
   {
     poly = new QEPolygonCellType(3);
     cellpointer.TakeOwnership(poly);
-    cellpointer->SetPointId(0, oddConnectivityCells[3 * i]);
+    cellpointer->SetPointId(0, oddConnectivityCells[static_cast<ptrdiff_t>(3 * i)]);
     cellpointer->SetPointId(1, oddConnectivityCells[3 * i + 1]);
     cellpointer->SetPointId(2, oddConnectivityCells[3 * i + 2]);
     mesh->SetCell(i, cellpointer);
